+++ Sources/Get/Request.swift
@@ -61,40 +61,6 @@ public struct Request<Response> {
     }
 }
 
-struct AnyEncodable: Encodable {
-    private let value: Encodable
-
-    init(_ value: Encodable) {
-        self.value = value
-    }
-
-    func encode(to encoder: Encoder) throws {
-        try value.encode(to: encoder)
-    }
-}
-
-extension URLRequest {
-    public func cURLDescription() -> String {
-        guard let url = url, let method = httpMethod else {
-            return "$ curl command generation failed"
-        }
-        var components = ["curl -v"]
-        components.append("-X \(method)")
-        for header in allHTTPHeaderFields ?? [:] {
-            let escapedValue = header.value.replacingOccurrences(of: "\"", with: "\\\"")
-            components.append("-H \"\(header.key): \(escapedValue)\"")
-        }
-        if let httpBodyData = httpBody {
-            let httpBody = String(decoding: httpBodyData, as: UTF8.self)
-            var escapedBody = httpBody.replacingOccurrences(of: "\\\"", with: "\\\\\"")
-            escapedBody = escapedBody.replacingOccurrences(of: "\"", with: "\\\"")
-            components.append("-d \"\(escapedBody)\"")
-        }
-        components.append("\"\(url.absoluteString)\"")
-        return components.joined(separator: " \\\n\t")
-    }
-}
-
 public struct Response<T> {
     public let value: T
     /// Original response data.